requirements-dev.txt
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0
pre-commit>=3.0.0
tox>=4.0.0
lxml-stubs>=0.4.0
